requirements.txt
Click==7.0
cloudpickle==0.8.0
cycler==0.10.0
dask==1.1.3
decorator==4.3.2
kiwisolver==1.0.1
matplotlib==3.0.3
networkx==2.2
numpy==1.16.2
Pillow==5.4.1
pyparsing==2.3.1
PyQt5==5.12
PyQt5-sip==4.19.14
pyqt5-tools==5.11.3.1.4
PySimpleGUI==3.25.0
python-dateutil==2.8.0
python-dotenv==0.10.1
PyWavelets==1.0.2
scikit-image==0.14.2
scipy==1.2.1
six==1.12.0
toolz==0.9.0
